refactor: merge scanner logic with new loader
