+++ src/utils/compatible-require.ts
