+++ src/utils/index.ts
@@ -1,19 +1,4 @@
-import path from 'path';
-import fs from 'fs';
-import compatibleRequire from './compatible-require';
 import minimatch from 'minimatch';
-
-export const loadFile = async (filepath: string) => {
-  try {
-    const extname = path.extname(filepath);
-    if (!extname) {
-      return fs.readFileSync(filepath);
-    }
-    return compatibleRequire(filepath);
-  } catch (err) {
-    throw err;
-  }
-}
 
 export function getDefaultExtensions() {
   return Object.keys(require.extensions)
