+++ test/fixtures/app-koa-with-ts/src/config/config.default.ts
@@ -1,5 +1,3 @@
-
-
 export default {
-    name: 'artus'
-}
+  name: 'artus'
+};
