+++ test/fixtures/app-koa-with-ts/src/config/plugin.default.ts
@@ -2,4 +2,4 @@
 import path from 'path';
 export default [
     { name: 'redis', path: path.resolve(__dirname, '../redis-plugin') },
-]
+];
